added vehicle super class.
