+++ Assets/Scripts/Bullet.cs
@@ -1,12 +1,10 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
 
 public class Bullet : MonoBehaviour
 {
-    public float speed = 10f;
     public float damage = 10f;
     public Vector3 direction;
+    public float speed = 10f;
     public string targetTag;
     private float ttl = 3f;
 
@@ -25,50 +23,47 @@ public class Bullet : MonoBehaviour
 
     private void OnCollisionEnter2D(Collision2D collision)
     {
-        if (collision.gameObject.tag == targetTag)
+        if (!collision.gameObject.CompareTag(targetTag)) return;
+        Destroy(gameObject);
+        if (collision.gameObject.GetComponent<Soldier>())
         {
-            Destroy(gameObject);
-            if(collision.gameObject.GetComponent<Soldier>())
+            Soldier enemy = collision.gameObject.GetComponent<Soldier>();
+            enemy.health -= damage;
+            if (enemy.health <= 0)
             {
-                Soldier enemy = collision.gameObject.GetComponent<Soldier>();
-                enemy.health -= damage;
-                if (enemy.health <= 0)
-                {
-                    Destroy(collision.gameObject);
-                }
-            }
-            else if (collision.gameObject.GetComponent<Tank>())
-            {
-                Tank enemy = collision.gameObject.GetComponent<Tank>();
-                enemy.health -= damage;
-                if (enemy.health <= 0)
-                {
-                    Destroy(collision.gameObject);
-                }
-            }
-            else if (collision.gameObject.GetComponent<MissileLauncher>())
-            {
-                MissileLauncher enemy = collision.gameObject.GetComponent<MissileLauncher>();
-                enemy.health -= damage;
-                if (enemy.health <= 0)
-                {
-                    Destroy(collision.gameObject);
-                }
-            }
-            else if (collision.gameObject.GetComponent<Castle>())
-            {
-                Castle enemy = collision.gameObject.GetComponent<Castle>();
-                enemy.health -= damage;
-                if (enemy.health <= 0)
-                {
-                    Destroy(collision.gameObject);
-                }
-            }
-            else
-            {
-                Debug.LogError("UNKNOWN ENEMY");
+                Destroy(collision.gameObject);
             }
         }
+        else if (collision.gameObject.GetComponent<Tank>())
+        {
+            Tank enemy = collision.gameObject.GetComponent<Tank>();
+            enemy.health -= damage;
+            if (enemy.health <= 0)
+            {
+                Destroy(collision.gameObject);
+            }
+        }
+        else if (collision.gameObject.GetComponent<MissileLauncher>())
+        {
+            MissileLauncher enemy = collision.gameObject.GetComponent<MissileLauncher>();
+            enemy.health -= damage;
+            if (enemy.health <= 0)
+            {
+                Destroy(collision.gameObject);
+            }
+        }
+        else if (collision.gameObject.GetComponent<Castle>())
+        {
+            Castle enemy = collision.gameObject.GetComponent<Castle>();
+            enemy.health -= damage;
+            if (enemy.health <= 0)
+            {
+                Destroy(collision.gameObject);
+            }
+        }
+        else
+        {
+            Debug.LogError("UNKNOWN ENEMY");
+        }
     }
-
 }
