+++ ProjectSettings/ProjectVersion.txt
@@ -1,1 +1,1 @@
-m_EditorVersion: 2018.3.8f1
+m_EditorVersion: 2018.3.6f1
